refactor(extensions): move AuthProvider to src/extensions/auth/, rename ext-jwt → ext-auth-jwt
- Move src/extensions/interfaces/auth-provider.ts → src/extensions/auth/auth-provider.ts
- New src/extensions/auth/index.ts barrel (export type for the 5 interfaces)
- src/extensions/interfaces/index.ts shim re-exports from ../auth/
- Rename extensions/ext-jwt → extensions/ext-auth-jwt (@veryfront/ext-auth-jwt)
- Update root deno.json workspace + add ./extensions/auth export
