+++ deno.json
@@ -7,8 +7,8 @@
     "./extensions/ext-ai-anthropic",
     "./extensions/ext-ai-google",
     "./extensions/ext-ai-openai",
+    "./extensions/ext-auth-jwt",
     "./extensions/ext-babel",
-    "./extensions/ext-jwt",
     "./extensions/ext-mdx",
     "./extensions/ext-node-compat",
     "./extensions/ext-opentelemetry",
@@ -70,6 +70,7 @@
     "./extensions": "./src/extensions/index.ts",
     "./extensions/interfaces": "./src/extensions/interfaces/index.ts",
     "./extensions/ai": "./src/extensions/ai/index.ts",
+    "./extensions/auth": "./src/extensions/auth/index.ts",
     "./testing": "./src/testing/index.ts",
     "./testing/assert": "./src/testing/assert.ts",
     "./testing/bdd": "./src/testing/bdd.ts",
