+++ extensions/ext-auth-jwt/deno.json
@@ -1,5 +1,5 @@
 {
-  "name": "ext-jwt",
+  "name": "@veryfront/ext-auth-jwt",
   "version": "0.1.0",
   "exports": "./src/index.ts",
   "veryfront": {
@@ -11,6 +11,7 @@
   },
   "imports": {
     "jose": "npm:jose@5.9.6",
+    "veryfront/extensions/auth": "../../src/extensions/auth/index.ts",
     "@std/assert": "jsr:@std/assert@1",
     "@std/testing/bdd": "jsr:@std/testing@1/bdd"
   },
