+++ extensions/ext-auth-jwt/src/verify-session-jwt.ts
