+++ src/extensions/auth/index.ts
@@ -1,0 +1,13 @@
+/**
+ * Auth category barrel — AuthProvider contract and token shapes.
+ *
+ * @module extensions/auth
+ */
+
+export type {
+  AuthProvider,
+  SignOptions,
+  TokenHeader,
+  TokenPayload,
+  VerifyOptions,
+} from "./auth-provider.ts";
